getProperties: Check if key exists instead of value in case of error
This commit fixes a double bug in the MapiObject->getProperties
function. First it now checks if the key (propertyKey | PT_ERROR) exists
in the fetched _properties array instead of the value. And secondly it
returns the value for the key (propertyKey | PT_ERROR) instead of the
value for the key propertyKey.